requirements.txt
python-dotenv
selenium
sendgrid
webdriver_manager

